examples/foc: move threads related logic to a separate file
